# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST in Golang with Echo Framework",
        "contact": {
            "name": "fei song"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/songfei1983/go-api-server/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/keys": {
            "put": {
                "description": "create an pair of key value",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "create key value",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/keys/{key}": {
            "get": {
                "description": "get string by key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetKeyValue a key",
                "operationId": "get-value-by-key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.KeyValue"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.KeyValue": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "object"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}